requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
gspread>=6.0.0
google-auth>=2.0.0
